[wasm-gc] static_assert that subtyping depth fits into TypeDefinition
Follow-up to https://crrev.com/c/4626327,
commit a7def102b6f0dc46f509e9eccf8dbcef294aa6ec.

Bug: v8:7748
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4625454
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#88377}

+++ src/wasm/module-decoder-impl.h
@@ -712,10 +712,10 @@ class ModuleDecoderImpl : public Decoder {
         errorf("type %u: forward-declared supertype %u", i, explicit_super);
         continue;
       }
-      uint8_t depth = module->types[explicit_super].subtyping_depth + 1;
+      uint32_t depth = module->types[explicit_super].subtyping_depth + 1;
       module_->types[i].subtyping_depth = depth;
       DCHECK_GE(depth, 0);
-      if (depth > static_cast<int>(kV8MaxRttSubtypingDepth)) {
+      if (depth > kV8MaxRttSubtypingDepth) {
         errorf("type %u: subtyping depth is greater than allowed", i);
         continue;
       }
